feat(core,lax,ad,ffi): add unsigned integers (U32/U64), half-precision floats (BF16/F16), and complex number primitives
Three major type system extensions across all crates:

1. Unsigned integer types (U32, U64):
   - Add DType::U32/U64, Literal::U32/U64, Value constructors/accessors
   - Type promotion rules: U32+I32->I64, U64+I64->F64, U32+F32->F64
   - Unsigned wrapping arithmetic (wrapping_add/sub/mul, checked_div/rem)
   - i128 comparison for correct mixed signed/unsigned ordering
   - Bitwise ops (Not, PopCount, CLZ, And, Or, Xor) for U32/U64
   - FFI marshalling with roundtrip tests for scalar and tensor U32/U64
   - E-graph literal encoding as "u32:N"/"u64:N" symbols
   - Partial eval, trace, and conformance support (exact tolerance tier)
   - Proptest strategies for U32/U64

2. Half-precision floats (BF16, F16) via `half` crate:
   - Add DType::BF16/F16, Literal::BF16Bits/F16Bits (u16 storage)
   - Constructors from_bf16_f32/from_f16_f32, accessors as_bf16_f32/as_f16_f32
   - Serde roundtrip, precision loss, range overflow tests
   - Proptest strategies for BF16 roundtrip and F16 finite values

3. Complex number primitives (Complex, Conj, Real, Imag):
   - Full complex binary arithmetic (Add/Sub/Mul/Div) with broadcasting
   - Complex unary ops (Neg, Abs, Exp, Log, Sin, Cos, Sqrt, Tanh)
   - ReduceSum with separate real/imaginary accumulators
   - VJP/JVP rules for all four primitives
   - Shape/dtype inference in trace (Complex64/128 output)
   - Batching support for Conj/Real/Imag (unary) and Complex (binary)
   - Extensive tests: Euler's formula, log roundtrip, conj involution,
     batch operations, comparison error for complex

+++ crates/fj-lax/src/comparison.rs
@@ -10,7 +10,7 @@ use crate::type_promotion::compare_literals;
 pub(crate) fn eval_comparison(
     primitive: Primitive,
     inputs: &[Value],
-    int_cmp: impl Fn(i64, i64) -> bool,
+    int_cmp: impl Fn(i128, i128) -> bool,
     float_cmp: impl Fn(f64, f64) -> bool,
 ) -> Result<Value, EvalError> {
     if inputs.len() != 2 {
